Fix some issues with extension messaging for transient pages. - Calling sendResponse within an onMessage listener no longer throws an error. - Sending a message to a transient page that is only partially loaded no longer fails.
BUG=120531
TEST=no

Review URL: https://chromiumcodereview.appspot.com/9968068

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130417 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/renderer/resources/extensions/miscellaneous_bindings.js
@@ -163,7 +163,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
             var rv = requestEvent.dispatch(request, sender, responseCallback);
             responseCallbackPreserved =
                 rv && rv.results && rv.results.indexOf(true) > -1;
-            if (!responseCallbackPreserved) {
+            if (!responseCallbackPreserved && port) {
               // If they didn't access the response callback, they're not
               // going to send a response, so clean up the port immediately.
               port.destroy_();
